requirements.txt
watchdog>=4.0.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
mcp[cli]>=1.2.0
apscheduler>=3.10.0
requests>=2.31.0
openai>=1.30.0
